dddd

+++ nebula-engine/src/main/java/nebula/data/db/TypeNames.java
@@ -10,7 +10,7 @@ import nebula.lang.RawTypes;
  */
 public class TypeNames {
 
-	private Map<RawTypes, Map<Long, String>> weighted = new HashMap<RawTypes, Map<Long, String>>();
+	private Map<RawTypes, Map<Integer, String>> weighted = new HashMap<RawTypes, Map<Integer, String>>();
 	private Map<RawTypes, String> defaults = new HashMap<RawTypes, String>();
 
 	public String get(RawTypes typecode) {
@@ -19,11 +19,11 @@ public class TypeNames {
 		return result;
 	}
 
-	public String get(RawTypes typeCode, long size, int precision, int scale)  {
-		Map<Long, String> map = weighted.get( typeCode );
+	public String get(RawTypes typeCode, int size, int precision, int scale)  {
+		Map<Integer, String> map = weighted.get( typeCode );
 		if ( map!=null && map.size()>0 ) {
 			// iterate entries ordered by capacity to find first fit
-			for (Map.Entry<Long, String> entry: map.entrySet()) {
+			for (Map.Entry<Integer, String> entry: map.entrySet()) {
 				if ( size <= entry.getKey() ) {
 					return replace( entry.getValue(), size, precision, scale );
 				}
@@ -32,16 +32,16 @@ public class TypeNames {
 		return replace( get(typeCode), size, precision, scale );
 	}
 	
-	private static String replace(String type, long size, int precision, int scale) {
+	private static String replace(String type, int size, int precision, int scale) {
 		type = type.replaceFirst("$s", Integer.toString(scale) );
 		type = type.replaceFirst("$l", Long.toString(size) );
 		return type.replaceFirst("$p", Integer.toString(precision) );
 	}
 
-	public void put(RawTypes typecode, long capacity, String value) {
-		Map<Long, String> map = weighted.get( typecode );
+	public void put(RawTypes typecode, int capacity, String value) {
+		Map<Integer, String> map = weighted.get( typecode );
 		if (map == null) {// add new ordered map
-			map = new TreeMap<Long, String>();
+			map = new TreeMap<Integer, String>();
 			weighted.put( typecode, map );
 		}
 		map.put(capacity, value);
